moved renameIdInData to services
